disable default features for surf and reqwest
Does not provide feature flags for enabling respective tls implementations, this is because it adds noise to features and in my opinion makes it hard to integrate. Instead, specify surf or reqwest in your crate features with whatever features you want.

note that `twitch_oauth2` has a feature `surf_client_curl` that enables curl.
this crate does not have anything similar.